% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{createAbstractSimulationSettings}
\alias{createAbstractSimulationSettings}
\title{Create simulation settings specifying confounding levels}
\usage{
createAbstractSimulationSettings(
  n = 1000,
  nX = 8,
  hCensor = 0.1,
  aIntercept = log(0.5),
  confoundingAySd = 0.5,
  confoundingmYSd = 0.5,
  confoundingAymSd = 0.5,
  mIntercept = log(0.1),
  mA = log(2),
  yIntercept = log(0.05),
  yA = log(0.5),
  yM = log(1.5)
)
}
\arguments{
\item{n}{Number of subject to simulate.}

\item{nX}{Number of covariates to simulate.}

\item{hCensor}{Hazard for random censoring.}

\item{aIntercept}{Intercept for probability of treatment (log scale).}

\item{confoundingAySd}{SD for the coefficients of the confounders between exposure and outcome.}

\item{confoundingmYSd}{SD for the coefficients of the confounders between mediator and outcome.}

\item{confoundingAymSd}{SD for the coefficients of the confounders between exposure, mediator, and outcome.}

\item{mIntercept}{Intercept for hazard of mediator (log scale)}

\item{mA}{Coefficient for hazard of mediator for the treatment. (log
scale).}

\item{yIntercept}{Intercept for hazard of outcome. (log scale)}

\item{yA}{Coefficient for the hazard of the outcome for treatment.
(log scale).}

\item{yM}{Coefficeint for the hazard of the outcome for the mediator.
(log scale).}
}
\value{
An abstract simulation settings object.
}
\description{
Create simulation settings specifying confounding levels
}
