% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RealWorldDataFunctions.R
\name{createMediatorRiskScore}
\alias{createMediatorRiskScore}
\title{Create the mediator risk score}
\usage{
createMediatorRiskScore(
  cohortMethodData,
  mediatorId,
  removeDuplicateSubjects = "keep all",
  riskWindowStart = 0,
  startAnchor = "cohort start",
  riskWindowEnd = 0,
  endAnchor = "cohort end",
  prior = createPrior("laplace", exclude = c(0), useCrossValidation = TRUE),
  control = createControl(noiseLevel = "silent", cvType = "auto", seed = 1,
    resetCoefficients = TRUE, tolerance = 2e-07, cvRepetitions = 1, fold = 10,
    startingVariance = 0.01)
)
}
\arguments{
\item{cohortMethodData}{An object of type \code{CohortMethodData} as generated using \code{CohortMethod::getDbCohortMethodData()}.}

\item{mediatorId}{The outcome ID corresponding to the mediator.}

\item{prior}{The prior used to fit the model. See \code{Cyclops::createPrior()} for details.}

\item{control}{The control object used to control the cross-validation used to determine the hyperparameters
of the prior (if applicable). See \code{Cyclops::createControl()} for details.}
}
\value{
A tibble with the mediator risk score (MRS).
}
\description{
Create the mediator risk score
}
