% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFitting.R
\name{createModelsettings}
\alias{createModelsettings}
\title{Create model fitting settings}
\usage{
createModelsettings(
  ps = "oracle",
  mrs = "oracle",
  psAdjustment = "matching",
  mrsAdjustment = "model",
  mediatorType = "time-to-event"
)
}
\arguments{
\item{ps}{Source of the propensity score. Can be "oracle" or "fit".}

\item{mrs}{Source of the mediator risk score. Can be "oracle" or "fit".}

\item{psAdjustment}{Type of PS adjustment. Can be "matching", "model",
"2D matching", "covariates", or "none".}

\item{mrsAdjustment}{Type of MRS adjustment. Can be "model", "2D matching",
"covariates", or "none".}

\item{mediatorType}{Type of mediator in the model. Can be "time-to-event" or
"binary".}
}
\value{
A tibble with the point estimates and 95 percent confidence intervals for the
main effect and mediator effect. The main effect is estimated both in a model
with and without the mediator.
}
\description{
Create model fitting settings
}
