% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{createSimulationSettings}
\alias{createSimulationSettings}
\title{Create simulation settings}
\usage{
createSimulationSettings(
  n = 1000,
  nX = 8,
  hCensor = 0.1,
  aIntercept = log(0.5),
  aX = c(1, -1, 1, -1, 0, 0, 0, 0),
  mIntercept = log(0.1),
  mX = c(0, 0, 0.5, -1, 1.5, -1, 0, 0),
  mA = log(2),
  yIntercept = log(0.05),
  yX = c(0.5, 0, 0.5, 0, -1, 0, 0, 0),
  yA = log(0.5),
  yM = log(1.5)
)
}
\arguments{
\item{n}{Number of subject to simulate.}

\item{nX}{Number of covariates to simulate.}

\item{hCensor}{Hazard for random censoring.}

\item{aIntercept}{Intercept for probability of treatment (log scale).}

\item{aX}{Coefficients for probability of treatment for the covariates.
(log scale).}

\item{mIntercept}{Intercept for hazard of mediator (log scale)}

\item{mX}{Coefficients for hazard of mediator for the covariates.
(log scale).}

\item{mA}{Coefficient for hazard of mediator for the treatment. (log
scale).}

\item{yIntercept}{Intercept for hazard of outcome. (log scale)}

\item{yX}{Coefficients for hazard of outcome for the covariates. (log
scale).}

\item{yA}{Coefficient for the hazard of the outcome for treatment.
(log scale).}

\item{yM}{Coefficeint for the hazard of the outcome for the mediator.
(log scale).}
}
\value{
A simulation settings object.
}
\description{
Create simulation settings
}
