% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RealWorldDataFunctions.R
\name{fitMediatorModel}
\alias{fitMediatorModel}
\title{Fit mediator model using real-world data}
\usage{
fitMediatorModel(
  studyPopulation,
  ps,
  mrs,
  psAdjustment = "matching",
  mrsAdjustment = "model",
  mediatorType = "time-to-event"
)
}
\arguments{
\item{studyPopulation}{The study population as created using the \code{CohortMethod::createStudyPopulation()}
function, with \code{outcomeId} specifying the outcome of interest.}

\item{ps}{The propensity score object as created using the \code{CohortMethod::createPs()} function.}

\item{mrs}{The mediator risk score object as created using the \code{createMrs()} function.}

\item{psAdjustment}{Type of PS adjustment. Can be "matching", "model",
"2D matching", or "none".}

\item{mrsAdjustment}{Type of MRS adjustment. Can be "model", "2D matching",
or "none".}

\item{mediatorType}{Type of mediator in the model. Can be "time-to-event" or
"binary".}
}
\value{
Returns a data frame with one row, and columns for the various model estimands.
}
\description{
Fit mediator model using real-world data
}
