% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MediationAnalysis.R
\docType{package}
\name{MediationAnalysis-package}
\alias{MediationAnalysis}
\alias{MediationAnalysis-package}
\title{MediationAnalysis: Performing a Mediation Analysis in Observational Data}
\description{
Code for evaluating various approaches to mediation analysis in observational data.
}
\author{
\strong{Maintainer}: Martijn Schuemie \email{schuemie@ohdsi.org}

}
\keyword{internal}
