% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RealWorldDataFunctions.R
\name{plotMrs}
\alias{plotMrs}
\title{Plot the mediator risk score}
\usage{
plotMrs(
  mrs,
  targetLabel = "Target",
  comparatorLabel = "Comparator",
  showFraction = TRUE
)
}
\arguments{
\item{mrs}{A tibble containing the MRS as created by \code{createMrs()}.}

\item{targetLabel}{A label to us for the target cohort.}

\item{comparatorLabel}{A label to us for the comparator cohort.}

\item{showFraction}{Add a label to the plot showing what fraction of the population has the
mediator during the time-at-risk?}
}
\value{
A ggplot object. Use the \code{ggplot2::ggsave()} function to save to file in a different format.
}
\description{
Plot the mediator risk score
}
