% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RealWorldDataFunctions.R
\name{plotMrsByExposure}
\alias{plotMrsByExposure}
\title{Plot the mediator risk score by exposure}
\usage{
plotMrsByExposure(
  mrs,
  targetLabel = "Target",
  comparatorLabel = "Comparator",
  showFraction = TRUE,
  title = NULL,
  fileName = NULL
)
}
\arguments{
\item{mrs}{A tibble containing the MRS as created by \code{createMrs()}.}

\item{targetLabel}{A label to us for the target cohort.}

\item{comparatorLabel}{A label to us for the comparator cohort.}

\item{showFraction}{Add a label to the plot showing what fraction of the population has the
mediator during the time-at-risk?}

\item{title}{Optional: the main title for the plot.}

\item{fileName}{Name of the file where the plot should be saved, for example 'plot.png'.
see the function \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}} for supported file formats.}
}
\value{
A ggplot object. Use the \code{ggplot2::ggsave()} function to save to file in a different format.
}
\description{
Plot the mediator risk score by exposure
}
