% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationLoop.R
\name{prepareForShinyApp}
\alias{prepareForShinyApp}
\title{Prepare simulation results for Shiny app}
\usage{
prepareForShinyApp(folder)
}
\arguments{
\item{folder}{Folder containing the simulation results.}
}
\value{
Does not return anything. Is executed for the side-effect of loading the results
into the inst/shinyApps/MediationResultsExplorer/data folder.
}
\description{
Prepare simulation results for Shiny app
}
