% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationLoop.R
\name{runSetOfSimulations}
\alias{runSetOfSimulations}
\title{Run a set of simulations}
\usage{
runSetOfSimulations(
  folder,
  simulationSettingsList,
  modelSettingsList,
  nSimulations = 1000,
  maxCores = 4
)
}
\arguments{
\item{folder}{Folder to write (intermediate) results.}

\item{simulationSettingsList}{A list of simulation settings objects as created
by \code{createSimulationSettings()}.}

\item{modelSettingsList}{A list of model fitting settings as created by
\code{createModelSettings()}.}

\item{nSimulations}{Number of times to repeat each simulation}

\item{maxCores}{Maximum number of CPU cores to use at the same
time.}
}
\value{
Does not return anything. Writes results to the folder.
}
\description{
Run a set of simulations
}
