% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{simulateData}
\alias{simulateData}
\title{Simulate data}
\usage{
simulateData(settings)
}
\arguments{
\item{settings}{A simulation settings object as created using \code{createSimulationSettings()} or
\code{createAbstractSimulationSettings()}}
}
\value{
A tibble with one row per subject and the following columns:
\itemize{
\item a: Was the patient treated (i.e. in the target group)?
\item tStart: Start of follow-up time (currently always 0).
\item tEnd: End of follow-up time (time to either outcome or random censor)
\item m: Did the patient experience the mediator during follow-up?
\item y: Did the patient experience the outcome at the end of follow-up?
\item tM: Time to mediator (truncated to t).
\item tY: Time to outcome (truncated to t).
\item pA: Probability of treatment.
\item hM_star: The hazard of the mediator, based on the counterfactual model (a
model without the exposure status).
\item x1 ... xn: The covariates.
}
}
\description{
Simulate data
}
