% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFitting.R
\name{createBootstrapSettings}
\alias{createBootstrapSettings}
\title{Create bootstrap settings}
\usage{
createBootstrapSettings(
  sampleSize = 1000,
  sampling = "strata",
  bootstrapType = "percentile",
  dropUninformativeStrata = TRUE
)
}
\arguments{
\item{sampling}{'strata' or 'person'}

\item{bootstrapType}{'reduced bias-corrected', 'bias-corrected', 'percentile' or 'pivoted'}

\item{dropUninformativeStrata}{TRUE or FALSE}

\item{sampleSize.}{Size of the bootstrap}
}
\description{
Create bootstrap settings
}
