% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFitting.R
\name{fitModel}
\alias{fitModel}
\title{Fit model to estimate mediator and main effect}
\usage{
fitModel(data, settings)
}
\arguments{
\item{data}{A data tibble as generated by the \code{simulateData()}
function.}

\item{settings}{A model settings object as created by the
\code{createModelsettings()} function.}
}
\value{
A tibble with one row and the following columns:
\itemize{
\item mainLogRr: The log of the hazard ratio for the main effect (effect of
treatment on outcome) in a model including the mediator.
\item mainLogLb: The log of the lower bound of the confidence interval for the
main effect in a model including the mediator.
\item mainLogUb: The log of the upper bound of the confidence interval for the
main effect in a model including the mediator.
\item mediatorLogRr: The log of the hazard ratio for the mediator effect
\item mediatorLogLb: The log of the lower bound of the confidence interval for the
mediator effect.
\item mediatorLogUb: The log of the upper bound of the confidence interval for the
mediator effect.
\item mainLogRrNoM: The log of the hazard ratio for the main effect (effect of
treatment on outcome) in a model \emph{not} including the mediator.
\item mainLogLbNoM: The log of the lower bound of the confidence interval for the
main effect in a model \emph{not} including the mediator.
\item mainLogUbNoM: The log of the upper bound of the confidence interval for the
main effect in a model \emph{not} including the mediator.
}
}
\description{
Fit model to estimate mediator and main effect
}
