% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RealWorldDataFunctions.R
\name{plotMrs}
\alias{plotMrs}
\title{Plot the mediator risk score}
\usage{
plotMrs(mrs, targetLabel = "Target", comparatorLabel = "Comparator")
}
\arguments{
\item{mrs}{A tibble containing the MRS as created by \code{createMrs()}.}

\item{targetLabel}{A label to us for the target cohort.}

\item{comparatorLabel}{A label to us for the comparator cohort.}
}
\value{
A ggplot object. Use the \code{ggplot2::ggsave()} function to save to file in a different format.
}
\description{
Plot the mediator risk score
}
